+++ tests/partition_tests.rs
@@ -7,7 +7,7 @@ use cuttana::stream::VertexStream;
 fn test_cuttana() {
     let data = vec![(0, vec![1, 2]), (1, vec![0]), (2, vec![0])];
     let stream = VertexStream::from_adjacency_list(data);
-    const NUM_PARTITIONS: u8 = 16;
+    const NUM_PARTITIONS: usize = 16;
     let config = CuttanaConfig::default();
 
     let result: PartitionResult<i32> = cuttana_partition(stream, NUM_PARTITIONS, config);
